backend/pyproject.toml
[tool.poetry]
name = "core"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.8"
aniso8601 = "8.0.0"
astroid = "2.4.2"
asyncpg = "^0.29.0"
attrs = "19.3.0"
certifi = "2020.4.5.1"
cffi = "^1.17.0"
chardet = "3.0.4"
click = "7.1.2"
dnspython = "1.16.0"
eventlet = "0.27.0"
fastavro = "0.23.4"
flask-cors = "3.0.8"
flask-executor = "0.9.3"
flask-jwt = "0.3.2"
flask-oidc = "1.4.0"
flask-restful = "0.3.8"
flask-restx = "1.3.0"
flask-sqlalchemy = "2.4.4"
flask = "1.1.4"
gevent = "^22.0.0"
greenlet = "^1.1.3"
gunicorn = "20.0.4"
httplib2 = "0.18.1"
idna = "2.9"
isort = "4.3.21"
itsdangerous = "1.1.0"
jinja2 = "2.11.2"
jsonschema = "3.2.0"
lazy-object-proxy = "1.4.3"
markupsafe = "1.1.1"
mccabe = "0.6.1"
minio = "7.1.2"
monotonic = "1.5"
neo4j = "4.0.2"
neobolt = "1.7.17"
neotime = "1.7.4"
numpy = "^1.18.5"
oauth2client = "4.1.3"
opentelemetry-exporter-jaeger = "1.6.2"
opentelemetry-instrumentation-flask = "0.26b1"
opentelemetry-instrumentation-requests = "0.26b1"
opentelemetry-instrumentation = "0.26b1"
psycopg2-binary = "2.9.9"
pyasn1-modules = "0.2.8"
pyasn1 = "0.4.8"
pycparser = "2.20"
pydantic = "1.8.2"
pyjwt = "1.4.2"
pylint = "2.5.3"
pyrsistent = "0.16.0"
python-dateutil = "2.8.1"
python-dotenv = "0.19.1"
python-json-logger = "0.1.11"
python-ldap = "3.3.1"
pytz = "2020.1"
requests = "2.26.0"
rsa = "4.6"
six = "1.15.0"
sqlalchemy = "1.3.17"
toml = "0.10.1"
tornado = "6.0.4"
urllib3 = "1.25.9"
werkzeug = "1.0.1"
wrapt = "1.12.1"
zope-event = "4.4"
zope-interface = "5.1.0"
common = { git = "https://github.com/vre-charite-dev/service_common.git", branch="feature/build-common-package" }
uvicorn = "^0.30.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
